Refactoring `cert-create` and removing redundant methods
